+++ DotNet/DataAcquisition/Listeners/DataAcquisitionRequestedListener.cs
@@ -58,7 +58,7 @@ public class DataAcquisitionRequestedListener : BaseListener<DataAcquisitionRequ
             throw new DeadLetterException("FacilityId is missing from the message key.", ex);
         }
 
-        var scope = _serviceScopeFactory.CreateScope();
+        using var scope = _serviceScopeFactory.CreateScope();
         var patientDataService =
             scope.ServiceProvider.GetRequiredService<IPatientDataService>();
