+++ DotNet/ServiceTests/IntegrationTests/Report/EndOfReportPeriodJobTests.cs
@@ -4,7 +4,6 @@ using LantanaGroup.Link.Report.Application.Options;
 using LantanaGroup.Link.Report.Core;
 using LantanaGroup.Link.Report.Domain;
 using LantanaGroup.Link.Report.Domain.Enums;
-using LantanaGroup.Link.Report.Domain.Managers;
 using LantanaGroup.Link.Report.Entities;
 using LantanaGroup.Link.Report.Entities.Enums;
 using LantanaGroup.Link.Report.Jobs;
@@ -20,6 +19,8 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Moq;
 using Quartz;
+using Quartz.Impl;
+using Quartz.Impl.Matchers;
 using Xunit.Abstractions;
 using Task = System.Threading.Tasks.Task;
 
@@ -41,14 +42,13 @@ namespace IntegrationTests.Report
         [Fact]
         public async Task Execute_AllReady_CallsManifestProducerAndUpdatesStatus()
         {
-            using var scope = _fixture.ServiceProvider.CreateScope();
-            var database = scope.ServiceProvider.GetRequiredService<IDatabase>();
-            var serviceScopeFactory = scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
-            var aggregator = scope.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
-            var dataAcqProducer = scope.ServiceProvider.GetRequiredService<DataAcquisitionRequestedProducer>();
-            var readyValProducer = scope.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
-            var auditProducer = scope.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
-            var tenantApiService = scope.ServiceProvider.GetRequiredService<ITenantApiService>();
+            var database = _fixture.ServiceProvider.GetRequiredService<IDatabase>();
+            var serviceScopeFactory = _fixture.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
+            var aggregator = _fixture.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
+            var dataAcqProducer = _fixture.ServiceProvider.GetRequiredService<DataAcquisitionRequestedProducer>();
+            var readyValProducer = _fixture.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
+            var auditProducer = _fixture.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
+            var tenantApiService = _fixture.ServiceProvider.GetRequiredService<ITenantApiService>();
 
             // Setup schedule
             var schedule = new ReportSchedule
@@ -106,7 +106,7 @@ namespace IntegrationTests.Report
             var submitKafkaMock = new Mock<IProducer<SubmitPayloadKey, SubmitPayloadValue>>();
             var submitPayloadProducer = new SubmitPayloadProducer(serviceScopeFactory, submitKafkaMock.Object);
 
-            var manifestProducerLogger = scope.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
+            var manifestProducerLogger = _fixture.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
             var manifestProducer = new ReportManifestProducer(manifestProducerLogger, serviceScopeFactory, aggregator, tenantApiMock.Object, blobStorageMock.Object, submitPayloadProducer, auditProducer);
 
             // Job context - FIXED
@@ -139,6 +139,8 @@ namespace IntegrationTests.Report
             await job.Execute(contextMock.Object);
 
             // Asserts
+            database = serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<IDatabase>();
+
             var updatedSchedule = await database.ReportScheduledRepository.SingleOrDefaultAsync(s => s.Id == schedule.Id);
             Assert.Equal(ScheduleStatus.EndOfPeriod, updatedSchedule.Status);
             Assert.True(updatedSchedule.EndOfReportPeriodJobHasRun);
@@ -151,15 +153,14 @@ namespace IntegrationTests.Report
         [Fact]
         public async Task Execute_NotAllReady_PatientsToEvaluate_CallsDataAcqProducer()
         {
-            using var scope = _fixture.ServiceProvider.CreateScope();
-            var database = scope.ServiceProvider.GetRequiredService<IDatabase>();
-            var serviceScopeFactory = scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
-            var aggregator = scope.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
-            var blobStorageService = scope.ServiceProvider.GetRequiredService<BlobStorageService>();
-            var submitPayloadProducer = scope.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
-            var readyValProducer = scope.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
-            var auditProducer = scope.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
-            var tenantApiService = scope.ServiceProvider.GetRequiredService<ITenantApiService>();
+            var database = _fixture.ServiceProvider.GetRequiredService<IDatabase>();
+            var serviceScopeFactory = _fixture.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
+            var aggregator = _fixture.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
+            var blobStorageService = _fixture.ServiceProvider.GetRequiredService<BlobStorageService>();
+            var submitPayloadProducer = _fixture.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
+            var readyValProducer = _fixture.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
+            var auditProducer = _fixture.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
+            var tenantApiService = _fixture.ServiceProvider.GetRequiredService<ITenantApiService>();
 
             // Setup schedule
             var schedule = new ReportSchedule
@@ -201,7 +202,7 @@ namespace IntegrationTests.Report
             var dataAcqKafkaMock = new Mock<IProducer<string, DataAcquisitionRequestedValue>>();
             var dataAcqProducer = new DataAcquisitionRequestedProducer(serviceScopeFactory, dataAcqKafkaMock.Object);
 
-            var manifestProducerLogger = scope.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
+            var manifestProducerLogger = _fixture.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
             var manifestProducer = new ReportManifestProducer(manifestProducerLogger, serviceScopeFactory, aggregator, tenantApiService, blobStorageService, submitPayloadProducer, auditProducer);
 
             // Job context
@@ -236,6 +237,8 @@ namespace IntegrationTests.Report
             // Asserts
             dataAcqKafkaMock.Verify(p => p.Produce(It.IsAny<string>(), It.IsAny<Message<string, DataAcquisitionRequestedValue>>(), null), Times.Exactly(pendingPatients.Count));
 
+
+            database = serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<IDatabase>();
             var updatedSchedule = await database.ReportScheduledRepository.SingleOrDefaultAsync(s => s.Id == schedule.Id);
             Assert.Equal(ScheduleStatus.EndOfPeriod, updatedSchedule.Status);
             Assert.True(updatedSchedule.EndOfReportPeriodJobHasRun);
@@ -246,15 +249,14 @@ namespace IntegrationTests.Report
         [Fact]
         public async Task Execute_NotAllReady_NeedsValidation_CallsReadyForValidationProducer()
         {
-            using var scope = _fixture.ServiceProvider.CreateScope();
-            var database = scope.ServiceProvider.GetRequiredService<IDatabase>();
-            var serviceScopeFactory = scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
-            var aggregator = scope.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
-            var blobStorageService = scope.ServiceProvider.GetRequiredService<BlobStorageService>();
-            var submitPayloadProducer = scope.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
-            var dataAcqProducer = scope.ServiceProvider.GetRequiredService<DataAcquisitionRequestedProducer>();
-            var auditProducer = scope.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
-            var tenantApiService = scope.ServiceProvider.GetRequiredService<ITenantApiService>();
+            var database = _fixture.ServiceProvider.GetRequiredService<IDatabase>();
+            var serviceScopeFactory = _fixture.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
+            var aggregator = _fixture.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
+            var blobStorageService = _fixture.ServiceProvider.GetRequiredService<BlobStorageService>();
+            var submitPayloadProducer = _fixture.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
+            var dataAcqProducer = _fixture.ServiceProvider.GetRequiredService<DataAcquisitionRequestedProducer>();
+            var auditProducer = _fixture.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
+            var tenantApiService = _fixture.ServiceProvider.GetRequiredService<ITenantApiService>();
 
             // Setup schedule
             var schedule = new ReportSchedule
@@ -294,9 +296,9 @@ namespace IntegrationTests.Report
             var readyValKafkaMock = new Mock<IProducer<ReadyForValidationKey, ReadyForValidationValue>>();
 
             // Create producer with mocked Kafka producer and resolved manager
-            var readyValProducer = new ReadyForValidationProducer(readyValKafkaMock.Object, scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>());
+            var readyValProducer = new ReadyForValidationProducer(readyValKafkaMock.Object, _fixture.ServiceProvider.GetRequiredService<IServiceScopeFactory>());
 
-            var manifestProducerLogger = scope.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
+            var manifestProducerLogger = _fixture.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
             var manifestProducer = new ReportManifestProducer(manifestProducerLogger, serviceScopeFactory, aggregator, tenantApiService, blobStorageService, submitPayloadProducer, auditProducer);
 
             // Job context
@@ -332,27 +334,27 @@ namespace IntegrationTests.Report
             readyValKafkaMock.Verify(p => p.Produce(It.Is<string>(topic => topic == nameof(KafkaTopic.ReadyForValidation)),
                 It.Is<Message<ReadyForValidationKey, ReadyForValidationValue>>(m => m.Key.FacilityId == schedule.FacilityId && m.Value.PatientId == "Patient1" && m.Value.ReportTrackingId == schedule.Id && m.Value.ReportTypes.Contains("TestReport") && m.Value.PayloadUri == "test://payload/patient1"), It.Is<Action<DeliveryReport<ReadyForValidationKey, ReadyForValidationValue>>>(h => h == null)), Times.Once());
 
+            database = serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<IDatabase>();
+
             var updatedSchedule = await database.ReportScheduledRepository.SingleOrDefaultAsync(s => s.Id == schedule.Id);
             Assert.Equal(ScheduleStatus.EndOfPeriod, updatedSchedule.Status);
             Assert.True(updatedSchedule.EndOfReportPeriodJobHasRun);
 
-            var asserDatabase = _fixture.ServiceProvider.GetRequiredService<IDatabase>();
-            var updatedEntry = await asserDatabase.SubmissionEntryRepository.FirstOrDefaultAsync(e => e.Id == entry.Id);
+            var updatedEntry = await database.SubmissionEntryRepository.FirstOrDefaultAsync(e => e.Id == entry.Id);
             Assert.Equal(ValidationStatus.Requested, updatedEntry.ValidationStatus);
         }
 
         [Fact]
         public async Task Execute_Exception_ReschedulesJob()
         {
-            using var scope = _fixture.ServiceProvider.CreateScope();
-            var database = scope.ServiceProvider.GetRequiredService<IDatabase>();
-            var serviceScopeFactory = scope.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
-            var aggregator = scope.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
-            var blobStorageService = scope.ServiceProvider.GetRequiredService<BlobStorageService>();
-            var submitPayloadProducer = scope.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
-            var readyValProducer = scope.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
-            var auditProducer = scope.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
-            var tenantApiService = scope.ServiceProvider.GetRequiredService<ITenantApiService>();
+            var database = _fixture.ServiceProvider.GetRequiredService<IDatabase>();
+            var serviceScopeFactory = _fixture.ServiceProvider.GetRequiredService<IServiceScopeFactory>();
+            var aggregator = _fixture.ServiceProvider.GetRequiredService<MeasureReportAggregator>();
+            var blobStorageService = _fixture.ServiceProvider.GetRequiredService<BlobStorageService>();
+            var submitPayloadProducer = _fixture.ServiceProvider.GetRequiredService<SubmitPayloadProducer>();
+            var readyValProducer = _fixture.ServiceProvider.GetRequiredService<ReadyForValidationProducer>();
+            var auditProducer = _fixture.ServiceProvider.GetRequiredService<AuditableEventOccurredProducer>();
+            var tenantApiService = _fixture.ServiceProvider.GetRequiredService<ITenantApiService>();
 
             // Setup schedule
             var schedule = new ReportSchedule
@@ -391,7 +393,7 @@ namespace IntegrationTests.Report
                 .Throws(new Exception("Test exception"));
             var dataAcqProducer = new DataAcquisitionRequestedProducer(serviceScopeFactory, dataAcqKafkaMock.Object);
 
-            var manifestProducerLogger = scope.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
+            var manifestProducerLogger = _fixture.ServiceProvider.GetRequiredService<ILogger<ReportManifestProducer>>();
             var manifestProducer = new ReportManifestProducer(manifestProducerLogger, serviceScopeFactory, aggregator, tenantApiService, blobStorageService, submitPayloadProducer, auditProducer);
 
             // Job context - FIXED
@@ -424,11 +426,134 @@ namespace IntegrationTests.Report
             await job.Execute(contextMock.Object);
 
             // Asserts
+            database = serviceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<IDatabase>();
+
             var updatedSchedule = await database.ReportScheduledRepository.SingleOrDefaultAsync(s => s.Id == schedule.Id);
             Assert.NotEqual(ScheduleStatus.EndOfPeriod, updatedSchedule.Status);
             Assert.False(updatedSchedule.EndOfReportPeriodJobHasRun);
 
             schedulerMock.Verify(s => s.DeleteJob(It.IsAny<JobKey>(), It.IsAny<CancellationToken>()), Times.Once());
+        }
+
+        [Fact]
+        public async Task MongoDb_And_InMemory_Schedulers_Are_Independent()
+        {
+            _fixture.ResetMocks();
+            //await _fixture.ClearDatabaseAsync();
+
+            Quartz.Logging.LogProvider.SetCurrentLogProvider(new NoOpLogProvider());
+            // This test verifies that the two keyed schedulers (MongoScheduler and InMemoryScheduler) 
+            // can coexist and operate independently
+
+            // Arrange: Create two separate in-memory schedulers to simulate the behavior
+            // Disable logging to avoid disposed LoggerFactory issues
+            ISchedulerFactory mongoSchedulerFactory = new StdSchedulerFactory(new System.Collections.Specialized.NameValueCollection
+            {
+                { "quartz.scheduler.instanceName", "MongoSimulatedScheduler" },
+                { "quartz.scheduler.exporter.type", "Quartz.Simpl.RemotingSchedulerExporter, Quartz" },
+                { "quartz.scheduler.exporter.bindName", "QuartzScheduler" },
+                { "quartz.scheduler.exporter.channelType", "tcp" },
+                { "quartz.serializer.type", "binary" }
+            });
+
+            ISchedulerFactory inMemorySchedulerFactory = new StdSchedulerFactory(new System.Collections.Specialized.NameValueCollection
+            {
+                { "quartz.scheduler.instanceName", "InMemorySimulatedScheduler" },
+                { "quartz.scheduler.exporter.type", "Quartz.Simpl.RemotingSchedulerExporter, Quartz" },
+                { "quartz.scheduler.exporter.bindName", "QuartzScheduler" },
+                { "quartz.scheduler.exporter.channelType", "tcp" },
+                { "quartz.serializer.type", "binary" }
+            });
+
+            IScheduler mongoScheduler = await mongoSchedulerFactory.GetScheduler();
+            IScheduler inMemoryScheduler = await inMemorySchedulerFactory.GetScheduler();
+
+            await mongoScheduler.Start(); // Start the schedulers
+            await inMemoryScheduler.Start();
+
+            await mongoScheduler.Clear();
+            await inMemoryScheduler.Clear();
+
+            // Define jobs for different schedulers
+            IJobDetail reportJob = JobBuilder.Create<DummyEndOfReportPeriodJob>()
+                .WithIdentity("EndOfReportPeriodJob", "ReportGroup")
+                .Build();
+
+            IJobDetail retryJob = JobBuilder.Create<DummyRetryJob>()
+                .WithIdentity("RetryJob", "RetryGroup")
+                .Build();
+
+            // Define triggers
+            ITrigger reportTrigger = TriggerBuilder.Create()
+                .WithIdentity("ReportTrigger", "ReportGroup")
+                .StartNow()
+                .WithDailyTimeIntervalSchedule(x => x.OnEveryDay().StartingDailyAt(Quartz.TimeOfDay.HourAndMinuteOfDay(1, 0)))
+                .ForJob(reportJob)
+                .Build();
+
+            ITrigger retryTrigger = TriggerBuilder.Create()
+                .WithIdentity("RetryTrigger", "RetryGroup")
+                .StartNow()
+                .WithSimpleSchedule(x => x.WithIntervalInMinutes(10).RepeatForever())
+                .ForJob(retryJob)
+                .Build();
+
+            // Schedule jobs on different schedulers
+            await mongoScheduler.ScheduleJob(reportJob, reportTrigger);
+            await inMemoryScheduler.ScheduleJob(retryJob, retryTrigger);
+
+            // Act: Query jobs from each scheduler
+            var mongoJobKeys = await mongoScheduler.GetJobKeys(GroupMatcher<JobKey>.AnyGroup());
+            var inMemoryJobKeys = await inMemoryScheduler.GetJobKeys(GroupMatcher<JobKey>.AnyGroup());
+
+            // Assert: Each scheduler has its own jobs
+            Assert.Single(mongoJobKeys);
+            Assert.Single(inMemoryJobKeys);
+
+            Assert.Contains(mongoJobKeys, k => k.Name == "EndOfReportPeriodJob");
+            Assert.Contains(inMemoryJobKeys, k => k.Name == "RetryJob");
+
+            // Verify they're truly independent
+            Assert.DoesNotContain(mongoJobKeys, k => k.Name == "RetryJob");
+            Assert.DoesNotContain(inMemoryJobKeys, k => k.Name == "EndOfReportPeriodJob");
+
+            // Verify scheduler instances are different
+            Assert.NotEqual(mongoScheduler.SchedulerName, inMemoryScheduler.SchedulerName);
+
+            // Cleanup
+            await mongoScheduler.Shutdown();
+            await inMemoryScheduler.Shutdown();
+        }
+    }
+
+    // Dummy job implementations for test purposes
+    public class DummyRetryJob : IJob
+    {
+        public Task Execute(IJobExecutionContext context)
+        {
+            // Simulate retry work
+            return Task.CompletedTask;
+        }
+    }
+
+    public class DummyEndOfReportPeriodJob : IJob
+    {
+        public Task Execute(IJobExecutionContext context)
+        {
+            // Simulate end of report period work
+            return Task.CompletedTask;
+        }
+    }
+
+    class NoOpLogProvider : Quartz.Logging.ILogProvider
+    {
+        public Quartz.Logging.Logger GetLogger(string name) => (level, func, exception, parameters) => true;
+        public IDisposable OpenNestedContext(string message) => new NoOpDisposable();
+        public IDisposable OpenMappedContext(string key, object value, bool destructure = false) => new NoOpDisposable();
+
+        private class NoOpDisposable : IDisposable
+        {
+            public void Dispose() { }
         }
     }
 }
